Read a ArcGIS FeatureLayer, Table, or ImageServer

arc_read() combines the functionality of arc_open() with arc_select() or arc_raster() to read an ArcGIS FeatureLayer, Table, or ImageServer to a sf object or object of class SpatRaster. Optionally set, check, or modify names for the returned data frame or sf object using the col_names and name_repair parameters.

Usage

arc_read(
  url,
  col_names = TRUE,
  col_select = NULL,
  n_max = getOption("arcgislayers.n_max", default = 10000),
  name_repair = "unique",
  crs = NULL,
  ...,
  fields = NULL,
  token = Sys.getenv("ARCGIS_TOKEN")
)

Arguments

url

The url of the remote resource. Must be of length one.

col_names

Default TRUE. If TRUE, use the default column names for the feature. If col_names is a character vector with the same length as the number of columns in the layer, the default names are replaced with the new names. If col_names has one fewer name than the default column names, the existing sf column name is retained. If col_names is the string "alias", names are set to match the available alias names for the layer.

col_select

Default NULL. A character vector of the field names to be returned. By default, all fields are returned.

n_max

Defaults to 10000 or an option set with options("arcgislayers.n_max" = <max records>). Maximum number of records to return.

name_repair

Default "unique". See vctrs::vec_as_names() for details. If name_repair = NULL, names are set directly.

crs

the spatial reference to be returned. If the CRS is different than the CRS for the input FeatureLayer, a transformation will occur server-side. Ignored if x is a Table.

...

Additional arguments passed to arc_select() if URL is a FeatureLayer or Table or arc_raster() if URL is an ImageLayer.

fields

a character vector of the field names that you wish to be returned. By default all fields are returned.

token

your authorization token. By default checks the environment variable ARCGIS_TOKEN. Set your token using arcgisutils::set_auth_token().

Value

An sf object, a data.frame, or an object of class SpatRaster.

Details

Examples

if (interactive()) {
 url <- "https://sampleserver6.arcgisonline.com/arcgis/rest/services/Census/MapServer/3"

 arc_read(url)

 # apply tolower() to column names
 arc_read(url, name_repair = tolower)

 url <- "https://sampleserver6.arcgisonline.com/arcgis/rest/services/EmergencyFacilities/FeatureServer/0"

 # use field aliases as column names
 arc_read(url, col_names = "alias")

 # read an ImageServer directly
 img_url <- "https://landsat2.arcgis.com/arcgis/rest/services/Landsat/MS/ImageServer"

 arc_read(
   img_url,
   width = 1000, height = 1000,
   xmin = -71, ymin = 43,
   xmax = -67, ymax = 47.5,
   bbox_crs = 4326
 )
}